Move `SpeculativeResources` to eagle_mab.py
